Move updateCodeLenses logic to AgentFixupControls
